Split editor logic into new module
