+++ editor.h
@@ -1,0 +1,10 @@
+#ifndef EDITOR_H
+#define EDITOR_H
+
+void run_editor(void);
+void set_up_term(void);
+void ee_init(void);
+void strings_init(void);
+void command_prompt(void);
+
+#endif /* EDITOR_H */
